fix(cc-widgets): move verifyCallback log method to incoming task test file
